1) java Date to kotlin Instant 2) Timber to Napier

+++ androidApp/src/main/kotlin/ru/samtakoy/features/notifications/learn_courses/UncompletedTaskApi.kt
@@ -146,7 +146,7 @@ class UncompletedTaskApi(
         cancelAllAlarmsAndNotifications()
 
         mPreferences.setUncompletedNotificationMinUtc(
-            DateUtils.getDateAfterCurrentLong(mS.shiftMillis)
+            DateUtils.getDateAfterCurrentLong(mS.shiftMillis.toLong())
         )
 
         planCheckingAlarm(mS.shiftMillis)
